reduced about windows to a simple form
