+++ mRemoteV1/packages.config
@@ -14,7 +14,6 @@
   <package id="DockPanelSuite.ThemeVS2015" version="3.0.6" targetFramework="net46" />
   <package id="Google.Protobuf" version="3.9.2" targetFramework="net472" />
   <package id="log4net" version="2.0.8" targetFramework="net46" />
-  <package id="Markdig" version="0.17.1" targetFramework="net472" />
   <package id="MySql.Data" version="8.0.17" targetFramework="net472" />
   <package id="NUnit3TestAdapter" version="3.15.1" targetFramework="net472" />
   <package id="ObjectListView.Official" version="2.9.1" targetFramework="net46" />
